refactor: extract HTTP setup from main
