+++ core/app_http.py
@@ -1,0 +1,37 @@
+from fastapi import FastAPI, Request
+from fastapi.middleware.cors import CORSMiddleware
+from fastapi.responses import JSONResponse
+from starlette.middleware.sessions import SessionMiddleware
+from uvicorn.middleware.proxy_headers import ProxyHeadersMiddleware
+
+from core.config import settings
+from core.logger import logger
+
+
+async def global_exception_handler(request: Request, exc: Exception):
+    logger.exception(f"Unhandled exception at {request.url}: {exc}")
+
+    if str(request.url.path).startswith("/admin"):
+        return JSONResponse(
+            status_code=500,
+            content={"error": "Internal Server Error", "detail": str(exc)},
+        )
+
+    return JSONResponse(
+        status_code=500,
+        content={"message": "Internal server error"},
+    )
+
+
+def configure_http(app: FastAPI) -> None:
+    app.exception_handler(Exception)(global_exception_handler)
+
+    app.add_middleware(ProxyHeadersMiddleware, trusted_hosts=["*"])
+    app.add_middleware(SessionMiddleware, secret_key=settings.SECRET_KEY)
+    app.add_middleware(
+        CORSMiddleware,
+        allow_origins=settings.CORS_ORIGINS,
+        allow_credentials=True,
+        allow_methods=["*"],
+        allow_headers=["*"],
+    )
